refactor: move user related models to users from topic
