+++ classes/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.admin.views.decorators import staff_member_required
 from django.contrib.auth.models import User
 
-from topic.models import Stage, Your_Stage, Initial_Password
+from users.models import Stage, Your_Stage, Initial_Password
 from topic.forms import StageEditForm, StudentForm
 # Create your views here.
 context = {}
